requirements.txt
# 基础依赖
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0

# 语音识别
SpeechRecognition==3.10.0
PyAudio==0.2.13

# 图像处理
Pillow==10.0.0
opencv-python==4.8.0.74
numpy==1.24.3

# 键盘鼠标控制
pyautogui==0.9.54 